#86 move replacements from contracts to utils
